+++ src/java/admin/controllers/DetailCourse.java
@@ -10,13 +10,11 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import org.hibernate.Session;
 import utility.Controller;
-import utility.Utils;
 
 /**
  *
@@ -31,7 +29,7 @@ public class DetailCourse extends Controller {
 
     @Override
     public void process(HttpServletRequest req, HttpServletResponse resp, Session session, HttpSession httpSession, PrintWriter out) throws Exception {
-        int courseId = Integer.parseInt(req.getParameter("courseId"));
+        long courseId = Long.parseLong(req.getParameter("courseId"));
 
         Course course = (Course) session.get(Course.class, courseId);
